refactor: Controller의 출력 OutputView로 분리
