+++ src/main/java/controller/AttendanceController.java
@@ -1,16 +1,13 @@
 package controller;
 
-import java.time.DayOfWeek;
 import java.time.LocalDate;
 import java.time.LocalTime;
-import java.time.format.TextStyle;
 import java.util.List;
-import java.util.Locale;
-import model.AbsentPenalty;
 import model.Attendance;
 import model.AttendanceBook;
 import model.CrewAttendances;
 import view.InputView;
+import view.OutputView;
 
 public class AttendanceController {
 
@@ -18,10 +15,12 @@ public class AttendanceController {
 
     private final LocalDate today;
     private final InputView inputView;
+    private final OutputView outputView;
 
-    public AttendanceController(LocalDate today, InputView inputView) {
+    public AttendanceController(LocalDate today, InputView inputView, OutputView outputView) {
         this.today = today;
         this.inputView = inputView;
+        this.outputView = outputView;
     }
 
     public void run() {
@@ -62,7 +61,7 @@ public class AttendanceController {
             }
 
             if (select.equals("4")) {
-                printRiskOfExpulsion(book);
+                printRiskOfDismissal(book);
                 continue;
             }
 
@@ -72,7 +71,6 @@ public class AttendanceController {
         }
     }
 
-    // 1번 기능
     private void attend(AttendanceBook book) {
         System.out.print(System.lineSeparator());
         String nickname = inputView.inputNickname();
@@ -81,35 +79,9 @@ public class AttendanceController {
         int minute = Integer.parseInt(time.split(":")[1]);
 
         Attendance attendance = book.check(nickname, today, LocalTime.of(hour, minute));
-        printAddInformation(attendance);
+        outputView.printAddInformation(attendance);
     }
 
-    // OutputView
-    private void printAddInformation(Attendance attendance) {
-        System.out.print(System.lineSeparator());
-        System.out.printf("%02d월 %02d일 %s %02d:%02d (%s)%n",
-                attendance.getMonth(), attendance.getDay(), getKoreanWeek(attendance.getDayOfWeek()),
-                attendance.getHour(), attendance.getMinute(), getAttendStatus(attendance));
-        System.out.print(System.lineSeparator());
-    }
-
-    private String getAttendStatus(Attendance attendance) {
-        if (attendance.isAbsent()) {
-            return "결석";
-        }
-
-        if (attendance.isLate()) {
-            return "지각";
-        }
-
-        return "출석";
-    }
-
-    private String getKoreanWeek(DayOfWeek dayOfWeek) {
-        return dayOfWeek.getDisplayName(TextStyle.FULL, Locale.KOREAN);
-    }
-
-    // 2번 기능
     private void update(AttendanceBook book) {
         System.out.print(System.lineSeparator());
         String updateNickname = inputView.inputUpdateNickname();
@@ -119,89 +91,21 @@ public class AttendanceController {
         Attendance attendance = book.findAttendance(updateNickname, updateDate);
         Attendance updateAttendance = book.update(updateNickname, updateDate, updateTime);
 
-        printUpdateInformation(attendance, updateAttendance);
+        outputView.printUpdateInformation(attendance, updateAttendance);
     }
 
-    // OutputView
-    private void printUpdateInformation(Attendance attendance, Attendance updateAttendance) {
-        System.out.print(System.lineSeparator());
-        System.out.printf("%02d월 %02d일 %s %02d:%02d (%s) -> ",
-                attendance.getMonth(), attendance.getDay(), getKoreanWeek(attendance.getDayOfWeek()),
-                attendance.getHour(), attendance.getMinute(), getAttendStatus(attendance));
-        System.out.printf("%02d:%02d (%s) 수정 완료!%n",
-                updateAttendance.getHour(), updateAttendance.getMinute(), getAttendStatus(updateAttendance));
-        System.out.print(System.lineSeparator());
-    }
-
-    // 3번 기능
     private void printAttendanceSheetsByCrew(AttendanceBook book) {
         System.out.print(System.lineSeparator());
         String nickname = inputView.inputNickname();
 
         CrewAttendances crewAttendances = book.findCrewAttendance(nickname);
-        printCrewAttendanceRecords(crewAttendances);
-    }
-
-    // OutputView
-    private void printCrewAttendanceRecords(CrewAttendances crewAttendances) {
-        System.out.print(System.lineSeparator());
-        System.out.printf("이번 달 %s의 출석 기록입니다.%n", crewAttendances.getNickname());
-        System.out.print(System.lineSeparator());
-
-        for (int day = 1; day < today.getDayOfMonth(); day++) {
-            LocalDate date = today.withDayOfMonth(day);
-            if (date.getDayOfWeek() == DayOfWeek.SATURDAY || date.getDayOfWeek() == DayOfWeek.SUNDAY
-            || date.isEqual(LocalDate.of(2024, 12, 25))) continue;
-
-            System.out.printf("%02d월 %02d일 %s ", date.getMonth().getValue(), day, getKoreanWeek(date.getDayOfWeek()));
-            if (crewAttendances.isAlreadyAttend(date)) {
-                Attendance attendance = crewAttendances.findAttendance(date);
-                System.out.printf("%02d:%02d (%s)%n", attendance.getHour(), attendance.getMinute(), getAttendStatus(attendance));
-                continue;
-            }
-
-            System.out.printf("--:-- (결석)%n");
-        }
-        System.out.print(System.lineSeparator());
-
-        System.out.printf("출석: %d회%n", crewAttendances.calculateAttendCountUntilDate(today));
-        System.out.printf("지각: %d회%n", crewAttendances.calculateLateCountUntilDate(today));
-        System.out.printf("결석: %d회%n", crewAttendances.calculateAbsentCountUntilDate(today));
-        System.out.print(System.lineSeparator());
-        AbsentPenalty absentPenalty = crewAttendances.determineAttendPenalty(today);
-        if (absentPenalty != AbsentPenalty.NONE) {
-            System.out.printf("%s 대상자입니다.%n", getAbsentPenalty(absentPenalty));
-        }
-        System.out.print(System.lineSeparator());
-    }
-
-    private String getAbsentPenalty(AbsentPenalty absentPenalty) {
-        if (absentPenalty == AbsentPenalty.DISMISSAL) {
-            return "제적";
-        }
-
-        if (absentPenalty == AbsentPenalty.INTERVIEW) {
-            return "면담";
-        }
-
-        if (absentPenalty == AbsentPenalty.WARNING) {
-            return "경고";
-        }
-
-        return "";
+        outputView.printCrewAttendanceRecords(crewAttendances, today);
     }
 
     // 4번 기능
-    private void printRiskOfExpulsion(AttendanceBook book) {
-        List<CrewAttendances> risk = book.findSortedRiskOfDismissalCrews(today);
+    private void printRiskOfDismissal(AttendanceBook book) {
+        List<CrewAttendances> riskOfDismissalCrews = book.findSortedRiskOfDismissalCrews(today);
 
-        System.out.print(System.lineSeparator());
-        System.out.println("제적 위험자 조회 결과");
-        for (CrewAttendances crew : risk) {
-            System.out.printf("- %s: 결석 %d회, 지각 %d회 (%s)%n",
-                    crew.getNickname(), crew.calculateAbsentCountUntilDate(today), crew.calculateLateCountUntilDate(today),
-                    getAbsentPenalty(crew.determineAttendPenalty(today)));
-        }
-        System.out.print(System.lineSeparator());
+        outputView.printRiskOfDismissal(riskOfDismissalCrews, today);
     }
 }
